model_api/requirements_docker.txt
opencv-python
numpy
onnxruntime-gpu
fastapi
uvicorn
requests
